moved libisosm library into libsm folder
