--- src/libisosm/Makefile.am
@@ -1,10 +1,0 @@
-EXTRA_DIST = Makefile.mak
-
-libisosm_la_SOURCES = iso-sm.c
-libisosm_la_CFLAGS = -I$(top_srcdir)/src
-
-noinst_LTLIBRARIES = libisosm.la
-
-noinst_HEADERS = \
-		 iso-sm-internal.h \
-		 iso-sm.h
