--- src/libisosm/Makefile.mak
@@ -1,11 +1,0 @@
-TOPDIR = ..\..
-
-TARGET = libisosm.lib
-OBJECTS = iso-sm.obj
-
-all: $(TARGET)
-
-$(TARGET): $(OBJECTS)
-        lib $(LIBFLAGS) /out:$(TARGET) $(OBJECTS)
-
-!INCLUDE $(TOPDIR)\win32\Make.rules.mak
