+++ src/libsceac/rw_sfid.c
@@ -21,7 +21,7 @@
 #include "config.h"
 #endif
 
-#include "libisosm/iso-sm.h"
+#include "libsm/iso-sm.h"
 #include "libopensc/internal.h"
 #include "libopensc/log.h"
 #include "rw_sfid.h"
