+++ src/libsm/iso-sm-internal.h
